requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
async-timeout==2.0.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
certifi==2018.1.18
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
daphne==1.4.2
Django==1.11.10
django-location-field==2.0.3
djangorestframework==3.7.7
httpie==0.9.9
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
msgpack-python==0.5.5
Pygments==2.2.0
pytz==2018.3
redis==2.10.6
requests==2.18.4
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
zope.interface==4.4.3
